python/v3/requirements.txt
# Core dependencies for Solar Heating System v3
# Hardware interfaces
megabas>=1.0.0
librtd>=1.0.0
lib4relind>=1.0.0

# MQTT communication
paho-mqtt>=1.6.1
asyncio-mqtt>=0.16.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
influxdb-client>=1.36.0

# TaskMaster AI integration
taskmaster-ai>=0.1.0
httpx>=0.25.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
